Modify : JS파일 위치  리팩토링
