+++ src/components/Nav/Nav.module.scss
@@ -64,7 +64,7 @@
     }
     .profileImg {
       width: 70px;
-      height: 80%;
+      height: 50px;
       border-radius: 30%;
       cursor: pointer;
       margin-right: 5px;
