+++ src/pages/Homepage/Comment.js
@@ -1,15 +1,17 @@
 import React, { useState } from 'react';
+import NestedComments from '../../pages/Homepage/NestedComments';
+import css from './Comment.module.scss';
 import BASE_URL from '../../config';
-import css from './Nestedcomment.module.scss';
-import NestedComments from '../NestedComments/NestedComments';
+import Nestedcomment from '../../pages/Homepage/Nestedcomment';
 
-function Nestedcomment({ data, pinId }) {
+function Comment({ data, pinId, nestedcomments }) {
   const [on, setOn] = useState(false);
   const [likeCount, setLikeCount] = useState(data.like_count);
   const now = new Date().getTime();
   const be = new Date(data.created_at);
   const before = be.getTime();
   const time = (now - before) / 1000 / 60;
+  const [nestOn, setNestOn] = useState(true);
 
   const NestedCommentsOn = () => {
     setOn(true);
@@ -54,6 +56,10 @@ function Nestedcomment({ data, pinId }) {
         }
       });
   };
+  const nestedcommentstoggle = () => {
+    setNestOn(prev => !prev);
+  };
+
   return (
     <div className={css.commentContainer}>
       <img className={css.commentImg} src={data.profile_image} alt="이미지" />
@@ -75,9 +81,23 @@ function Nestedcomment({ data, pinId }) {
           </div>
         </div>
         {on && <NestedComments data={data} setOn={setOn} pinId={pinId} />}
+        {nestedcomments[0] &&
+          (nestOn ? (
+            <div onClick={nestedcommentstoggle} className={css.nestedcomment}>
+              --- 답변 {nestedcomments.length}개 보기
+            </div>
+          ) : (
+            <div onClick={nestedcommentstoggle} className={css.nestedcomment}>
+              --- 답변 숨기기
+            </div>
+          ))}
+        {(nestedcomments[0] && nestOn) ||
+          nestedcomments.map(data => {
+            return <Nestedcomment key={data.id} data={data} pinId={pinId} />;
+          })}
       </div>
     </div>
   );
 }
 
-export default Nestedcomment;
+export default Comment;
