+++ src/pages/Homepage/Finfeedmodal.module.scss
@@ -92,9 +92,11 @@
             height: 90%;
             border-radius: 100%;
             margin-right: 10px;
+            cursor: pointer;
           }
           .userId {
             font-size: 20px;
+            cursor: pointer;
           }
           .follow {
             font-size: 15px;
